refactor(error-tracking): colocate status select in Status.tsx
